requirements.txt
schedule
teslapy
geopy>=1.20.0
requests-oauthlib>=1.3.0
websocket_client>=0.59.0
selenium>=3.141.0
pywebview>=3.2
